Add Support for Nested Packages in list
I created a couple of methods in dataone_register.py that contains code which was duplicated before. You'll see some removed lines with the new function call replacement in the harvester.

The changes in repository.py are documentation and a small pep8 suggestion (`params` wasn't being used).

Testing Files:
  I renamed the datone_test & associated json file to harvester_test, and then createds a new test for the dataone stuff (dataone_test.py).
The testing updates puts us at 89%-we'll see how circleci reports it on the web.