Moved JSON Response to External File, Dict Access Changes
I moved the nested json structure to a new file and named it to match the harvester test.

I switched dictionary access to using the get method-which allowed my to get rid of some of the try catch blocks.

I also switched the bogus urls to localhost_01 and dictionary asserts to assertDictEqual.